+++ skit_labels/utils.py
@@ -3,10 +3,12 @@ Module provides access to logger config, session token and package version.
 """
 import os
 import sys
-from typing import Optional
 
 import toml
+from typing import Optional
 from loguru import logger
+from datetime import datetime
+
 
 LOG_LEVELS = ["CRITICAL", "ERROR", "WARNING", "SUCCESS", "INFO", "DEBUG", "TRACE"]
 
@@ -68,3 +70,33 @@ def read_session() -> Optional[str]:
             return handle.read().strip()
     except FileNotFoundError:
         return None
+
+
+def to_datetime(d: str) -> datetime:
+    """
+    Convert an arbitrary date string to ISO format.
+
+    :param d: A date string.
+    :type d: str
+    :raises ValueError: If the date string can't be parsed.
+    :return: The date string in ISO format.
+    :rtype: datetime
+    """
+    if not isinstance(d, str):
+        raise TypeError(f"Expected a string, got {type(d)}")
+
+    time_fns = [
+        datetime.fromisoformat,
+        lambda date_string: datetime.strptime(
+            date_string, "%Y-%m-%d %H:%M:%S.%f %z %Z"
+        ),
+        lambda date_string: datetime.strptime(date_string, "%Y-%m-%dT%H:%M:%SZ"),
+        lambda date_string: datetime.strptime(date_string, "%Y-%m-%dT%H:%M:%S.%f%z"),
+    ]
+
+    for time_fn in time_fns:
+        try:
+            return time_fn(d)
+        except ValueError:
+            continue
+    raise ValueError(f"Could not parse date string: {d}")